projet-django/location_app/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
pillow
psycopg2-binary
